+++ srcs/Dispatch.cpp
@@ -1,0 +1,9 @@
+#include <Dispatch.hpp>
+#include <sys/epoll.h>
+
+Dispatch::Dispatch(void)
+: _epollfd(epoll_create1(0)) {
+  if (_epollfd == -1) {
+    throw serverException("Error: failed to create epoll instance.");
+  }
+}
